% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{ape}
\alias{ape}
\title{Average partial (i.e., marginal) total, direct and indirect effects from a pmlsprobit object}
\usage{
ape(mod, variables = NULL)
}
\arguments{
\item{mod}{Object of class inheriting from "pmlsbp"}

\item{variables}{a string including the variable names of the covariates for
which the marginal effects should be calculated using numeric derivatives.  It is valid only if \code{type="me"}.
This option should be used only if the model contains non-linear terms in the specification \eqn{X \beta}}
}
\description{
Average partial (i.e., marginal) total, direct and indirect effects from a pmlsprobit object
}
