% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmlspprobit-package.R
\docType{package}
\name{pmlsbp-package}
\alias{pmlsbp}
\alias{pmlsbp-package}
\title{pmlsbp: Estimate spatial autoregressive nonlinear probit models with autoregressive disturbances}
\description{
Estimate spatial autoregressive nonlinear probit models with autoregressive disturbances. The package allows the estimation of SAR and SARAR probit models using partial maximum likelihood estimation. Estimation and inference regarding marginal effects is also possible.
}
\keyword{internal}
