% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_probit.R
\name{pmlsbp}
\alias{pmlsbp}
\alias{summary.pmlsprobit}
\alias{vcov.pmlsprobit}
\alias{coef.pmlsprobit}
\alias{print.pmlsprobit}
\title{pmlsbp: Fitting Spatial Probit Models}
\usage{
pmlsbp(formula,data, model="SAR", grouping=2, W=NULL,
 zero.policy =spatialreg::get.ZeroPolicyOption(), M=NULL, formula_xlag=NULL,
 W2=NULL, y=NULL, X=NULL, method_inv="chol", start=NULL, subset=NULL,
 na.action=na.fail,qu=Inf,
 mvtnorm_control=list(M=25e3, sim_type="mc" , tol = .Machine$double.eps, fast = FALSE),
 finalHessian=TRUE, method="bfgsr",print.level=2, vce.type=c("asy") ,
 Conley=list(coords=NULL,LM=2),nBoot=1e3 , ...)

\method{summary}{pmlsprobit}(object, eigentol = 1e-12, ...)

\method{vcov}{pmlsprobit}(x)

\method{coef}{pmlsprobit}(x)

\method{print}{pmlsprobit}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{formula}{an object of class "formula": a symbolic description of the model to be fit. The details of model specification are given for lm()}

\item{data}{an optional data frame containing the variables in the model. By default the variables are taken from the environment which the function is called.}

\item{model}{a character (\code{"SAR"} or \code{"SARAR"}) defining the model to be estimated. The default is \code{"SAR"}.}

\item{grouping}{an integer defining the number of observations to include in the tuples
to be used for estimation, the default is 2.}

\item{W}{a listw object created for example by \code{nb2listw}}

\item{zero.policy}{default NULL, use global option value; if TRUE assign zero to the lagged
value of zones without neighbours, if FALSE (default) assign NA - causing
\code{pmlsbp} to terminate with an error
(see also \code{\link[spatialreg]{set.ZeroPolicyOption}}) and the same option
in \code{\link[spatialreg]{GMerrorsar}})}

\item{M}{a listw object created for example by \code{nb2listw} to be used
for the model with spatial autoregressive disturbances. This is relevant only if \code{model="SARAR"}.}

\item{formula_xlag}{an object of class "formula": a symbolic description of the variables to be spatially lagged}

\item{W2}{a listw object created for example by \code{nb2listw} to be used
to create the spatial weight matrix to multiply the covariates. This is relevant only if \code{formula_xlag} is not \code{NULL}.}

\item{method_inv}{a character for setting the inversion of \eqn{I-\rho*W}, either \code{"solve"}, \code{"chol"} or \code{"fast"}.
\code{"solve"} uses \code{solve}, "chol" performs the inversion of the matrix from its Choleski decomposition (see \code{\link[Matrix]{chol2inv}}), \code{"fast"} uses an approximation that requires option \code{qu}}

\item{start}{numeric vector, initial value of parameters.}

\item{subset}{an optional vector specifying a subset of observations to be used for fitting}

\item{na.action}{a function (default \code{na.fail}), can also be \code{na.omit} or \code{na.exclude}
with consequences for residuals and fitted values - in these cases the weights
list will be subsetted to remove NAs in the data. It may be necessary to set
\code{zero.policy} to TRUE because this subsetting may create no-neighbour observations.
Note that only weights lists created without using the \code{glist} argument to nb2listw may be subsetted.}

\item{qu}{an integer to be used only if \code{method_inv="fast"}}

\item{mvtnorm_control}{an optional object of class \code{"list"} that includes the controls for
the calculation of the multivariate normal probabilities.  The list members should include
\code{M}, \code{rmtol} and \code{fast} (see \code{\link[mvtnorm]{lmpvnorm}}) and
\code{simtype}, either \code{"mc"} for Monte-Carlo procedure or \code{"qmc"}, quasi-Monte-Carlo procedure)
The default is \code{list(M=25e3, sim_type="mc" , tol = .Machine$double.eps, fast = FALSE)}}

\item{finalHessian}{how (and if) to calculate the final Hessian. Either \code{FALSE} (not calculate), \code{TRUE}
(use analytic/numeric Hessian) or \code{"bhhh"/"BHHH"} for information equality approach
(see \code{\link[maxLik]{maxLik}}).}

\item{method}{maximisation method to be used by \code{maxLik}. The default is \code{method="bfgsr"}
Broyden-Fletcher-Goldfarb-Shanno as implemented in R. See also the same option in (see \code{\link[maxLik]{maxLik}}).}

\item{print.level}{Integer, debugging information. Larger number prints more details (see \code{\link[maxLik]{maxLik}}).}

\item{vce.type}{an optional character specifying the type of variance covariance
estimation  (VCE) to be performed. It can be either \code{"asy"} for the asymptotic VCE,
\code{"bootstrap"} for bootstrap  VCE or \code{"mConley"} for the modified Conley approach.
The default is "asy".}

\item{Conley}{a list with two members, which is relevant only for \code{vce.type="mConley"}.
The first member (\code{coords}) is a matrix with the coordinates of each observation;
the second member (\code{LM}) is the number of neighbors to be used.
The default is \code{list(coords=NULL,LM=2)}}

\item{nBoot}{an integer, the number of bootstrap samples to be used for vce.type="bootstrap".
It is relevant only if \code{vce.type="bootstrap"}.}

\item{...}{}
}
\value{
\code{pmlsbp} returns an object of class \code{"pmlsprobit"}.
The function \code{summary} is used to obtain and print a summary of the results.
A object of class \code{"pmlsprobit"} is list containing the following components:
\tabular{ll}{
\code{beta} \tab a named vector including the \eqn{\beta} and \eqn{\gamma}, the coefficents attached to the variables resulting from \code{formula} and  \code{formula_xlag} \cr
\code{call} \tab the call \cr
\code{code} \tab \cr
\code{estimate} \tab a named vector of model estimates \cr
\code{f} \tab a list including useful matrices to be used in \code{predict} \cr
\code{maximum} \tab the value of the partial log-likelihood given the data and the model estimates \cr
\code{message} \tab convergence message \cr
\code{model} \tab \cr
\code{model.type} \tab either \code{"SAR"} or  \code{"SARAR"}  \cr
\code{rho} \tab a named scalar for the \eqn{\rho} parameter \cr
\code{slx} \tab \code{TRUE} if the model includes spatially lagged covariates \cr
\code{start} \tab a named vector, the starting values used in estimation \cr
\code{terms} \tab 	the terms object used. \cr
\code{vcov} \tab the variance-covariance matrix of the estimates \cr
\code{W} \tab matrix \eqn{W}  \cr
\code{lambda} \tab a named scalar for parameter \eqn{lambda} (only for SARAR models)  \cr
\code{M} \tab matrix \eqn{M} (only for SARAR models)  \cr
\code{W2} \tab matrix \eqn{W_2} (only if \code{slx=TRUE})  \cr
}
}
\description{
Estimation of Partial Maximum Likelihood Spatial Autoregressive (SAR) Probit Models and Spatial Autoregressive Probit Models with Autoregressive Disturbances (SARAR).
The model structure for the SARAR model is: \cr
\tabular{c}{
\eqn{y^* = \rho W y + X \beta + W_2 Z \gamma} + u  \cr
\eqn{u = \lambda M u + \epsilon} \cr
}
where \eqn{\epsilon \sim N(0,\sigma^2)} and \eqn{y=1(y^*>0)}. The SAR model assumes \eqn{\lambda=0}
}
\section{Functions}{
\itemize{
\item \code{summary(pmlsprobit)}: computes and returns a list of summary statistics of the fitted spatial probit model

\item \code{vcov(pmlsprobit)}: returns the variance covariance matrix of a \code{"pmlsprobit"} object

\item \code{coef(pmlsprobit)}: returns the coefficients estimate from a \code{"pmlsprobit"} object

\item \code{print(pmlsprobit)}: prints the call and the coefficient estimates of a \code{"pmlsprobit"} object

}}
\references{
Bille', A. G., & Leorato, S. (2020). Partial ML estimation for spatial autoregressive nonlinear probit models with autoregressive disturbances. Econometric Reviews, 39(5), 437-475.
}
