% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.pmlsprobit}
\alias{predict.pmlsprobit}
\title{Predicted values based on pmlsprobit object}
\usage{
\method{predict}{pmlsprobit}(
  object,
  type = "prob",
  newdata = NULL,
  variables = NULL,
  delta.method = F,
  ...
)
}
\arguments{
\item{object}{a \code{"pmlsprobit"} object}

\item{type}{\code{"prob"} for the probability \eqn{P(Y_i=1)} or \code{"me"}
for the direct, indirect and total average marginal effect for each observation}

\item{newdata}{An optional data frame in which to look for variables with which to predict.
If omitted, the fitted values are used.}

\item{variables}{a string including the variable names of the covariates for
which the marginal effects should be calculated using numeric derivatives.  It is valid only if \code{type="me"}.
This option should be used only if the model contains non-linear terms in the specification \eqn{X \beta}.}

\item{delta.method}{a boolean. If \code{TRUE} the standard errors of the predictions are calculated.
It is valid only if \code{type="me"}}
}
\description{
Predicted values based on pmlsprobit object
}
